Factor out prvCheckOptions inner loop to function
The function prvCheckOptions_outerLoop had previously contained the body
of an outer loop, as well as an inner loop. This commit moves the inner
loop into its own function, prvCheckOptions_innerLoop. This is to enable
automated property checking of the inner loop separately from the rest
of the function.

The new function takes all variables that are accessed in the inner loop
by reference.  This commit does not make any whitespace changes;
indentation will be fixed in a later commit.